+++ dev/sparktestsupport/modules.py
@@ -600,7 +600,13 @@ pyspark_ml = Module(
         "pyspark.ml.tests.test_pipeline",
         "pyspark.ml.tests.test_stat",
         "pyspark.ml.tests.test_training_summary",
-        "pyspark.ml.tests.test_tuning",
+        "pyspark.ml.tests.tuning.test_tuning",
+        "pyspark.ml.tests.tuning.test_cv_io_basic",
+        "pyspark.ml.tests.tuning.test_cv_io_nested",
+        "pyspark.ml.tests.tuning.test_cv_io_pipeline",
+        "pyspark.ml.tests.tuning.test_tvs_io_basic",
+        "pyspark.ml.tests.tuning.test_tvs_io_nested",
+        "pyspark.ml.tests.tuning.test_tvs_io_pipeline",
         "pyspark.ml.tests.test_util",
         "pyspark.ml.tests.test_wrapper",
         "pyspark.ml.torch.tests.test_distributor",
